requirements.dev.txt
black==20.8b1
celluloid==0.2.0
flake8==3.9.2
ipympl==0.5.8
isort
jupyter==1.0.0
matplotlib==3.3.2
mypy==0.790
nbqa==0.12.0
prettytable==2.1.0
pytest==6.1.2
torchviz==0.0.2
